requirements.txt
Flask==2.0.2
tensorflow==2.8.0
Pillow==8.4.0
Werkzeug==2.0.2
numpy==1.21.2